SpringSecurity6 and jsonwebtoken implemented. User Authentication completed. Login and Register page created.

+++ NoteMakingApplication/src/main/java/com/DXCTrainingProject/NoteMakingApplication/Service/NoteServiceImpl.java
@@ -4,9 +4,13 @@ import ExHandler.NoteNotFoundException;
 import com.DXCTrainingProject.NoteMakingApplication.DTO.NoteDTO;
 import com.DXCTrainingProject.NoteMakingApplication.DTO.NoteResponseDTO;
 import com.DXCTrainingProject.NoteMakingApplication.Entity.Note;
+import com.DXCTrainingProject.NoteMakingApplication.Entity.User;
 import com.DXCTrainingProject.NoteMakingApplication.Repository.NoteRepository;
+import com.DXCTrainingProject.NoteMakingApplication.Repository.UserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,9 +24,25 @@ public class NoteServiceImpl implements NoteService {
     @Autowired
     private NoteRepository noteRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
+    private User getCurrentuser(){
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+        if(principal instanceof UserDetails){
+            username = ((UserDetails) principal).getUsername();
+        }else{
+            username = principal.toString();
+        }
+        return userRepository.findByUsername(username).orElseThrow(()-> new RuntimeException("Authenticated user not found"));
+    }
+
     @Override
     public NoteResponseDTO createNote(NoteDTO noteDTO) {
+        User user = getCurrentuser();
         Note note = modelMapper.map(noteDTO, Note.class);
+        note.setUser(user);
         Note savedNote = noteRepository.save(note);
         return modelMapper.map(savedNote, NoteResponseDTO.class);
     }
@@ -36,7 +56,13 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public NoteResponseDTO updateNote(Long noteId, NoteDTO noteDTO) {
+        User user = getCurrentuser();
         Note noteFromDb = noteRepository.findById(noteId).orElseThrow(() -> new NoteNotFoundException("Note not found"));
+
+        if(!noteFromDb.getUser().getId().equals(user.getId())){
+            throw new NoteNotFoundException("Note not found for this user");
+        }
+
         noteFromDb.setNoteTitle(noteDTO.getNoteTitle());
         noteFromDb.setNoteDescription(noteDTO.getNoteDescription());
         noteFromDb.setTag(noteDTO.getTag());
@@ -50,7 +76,14 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public NoteResponseDTO updateNoteStatus(Long noteId) {
+        User user = getCurrentuser();
+
         Note noteFromDb = noteRepository.findById(noteId).orElseThrow(() -> new NoteNotFoundException("Note not found"));
+
+        if(!noteFromDb.getUser().getId().equals(user.getId())){
+            throw new NoteNotFoundException("Note not found for this user");
+        }
+
         if (noteFromDb.getIsCompleted() == true) {
             noteFromDb.setIsCompleted(false);
         } else {
@@ -66,15 +99,32 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public String deleteNote(Long noteId) {
+        User user = getCurrentuser();
         Note note = noteRepository.findById(noteId).orElseThrow(() -> new NoteNotFoundException("Note with this id not found."));
+
+        if (!note.getUser().getId().equals(user.getId())) {
+            throw new NoteNotFoundException("Note not found for this user");
+        }
+
         noteRepository.delete(note);
         return "Note delete successfully.";
     }
 
     @Override
     public List<NoteResponseDTO> searchNotes(String keyword) {
-        List<Note> notes = noteRepository.searchNotes(keyword);
+        User user = getCurrentuser();
+        List<Note> notes = noteRepository.searchNotesByUser(user, keyword);
         List<NoteResponseDTO> notesResponseDTOS = notes.stream().map(note -> modelMapper.map(note, NoteResponseDTO.class)).toList();
         return notesResponseDTOS;
     }
+
+    @Override
+    public List<NoteResponseDTO> getAllNotesOfUser() {
+
+        User user = getCurrentuser();
+        List<Note> notes = noteRepository.findByUser(user);
+        return notes.stream().map(note -> modelMapper.map(note, NoteResponseDTO.class)).toList();
+    }
+
+
 }
